By default, add a secgroup rule to permit all private network traffic.
This is useful for debugging purposes.  It's often useful to access tenant VMs
via ICMP/TCP from within the router VM itself;  The default demo tenant
security group rules prevent this because they restrict ingress traffic to the
tenant's default security group (the router VM is on a different tenant, and
has a different security group).